Add Bitmask.fromBitmask constructor
This creates a Bitmask object that is a deep copy of another, whereas
Bitmask.operator = creates a second Bitmask variable that points to the
first one.

Used to fix bugs in the Bitmask operators where the current object was
being modified when it should have been returning a diiferent
object.